backend (3)/backend/requirements.txt
fastapi
flask-cors
pymongo
python-dotenv
passlib[bcrypt]
transformers
sentencepiece
xgboost
scikit-learn==1.6.1
joblib
pandas
numpy
gunicorn
pymongo
python-dotenv
python-multipart
openai
torch
uvicorn
groq
google-auth
requests
